\name{Data_Maize}
\alias{Data_Maize}
\docType{data}
\title{
Monthly International Maize Price Data
}
\description{
Monthly international Maize price (Dollor per million ton) from January 2001 to December 2019.
}
\usage{data("Data_Maize")}
\format{
  A time series data with 228 observations.
  \describe{
    \item{\code{price}}{a time series}
  }
}
\details{
Dataset contains 228 observations of monthly international Maize price (Dollor per million ton). It is obtained from World Bank "Pink sheet".
}
\source{
https://www.worldbank.org/en/research/commodity-markets
}
\references{
https://www.worldbank.org/en/research/commodity-markets
}
\examples{
data(Data_Maize)
}
\keyword{datasets}
